# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kasir Boy API",
        "description": "API Documentation for Kasir Boy Backend",
        "contact": {
            "email": "muhfawas7@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get all transactions with pagination",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by specific date (Y-m-d) - for single day filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-01-21"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter by start date (Y-m-d) - for date range",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter by end date (Y-m-d) - for date range",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-01-31"
                        }
                    },
                    {
                        "name": "payment_method",
                        "in": "query",
                        "description": "Filter by payment method",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cash",
                                "qris",
                                "debit"
                            ]
                        }
                    },
                    {
                        "name": "cashier_id",
                        "in": "query",
                        "description": "Filter by cashier ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by order number, cashier name, or cashier username",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "fawas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transactions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transactions retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "ab765dced58cfe2f620929db31452145",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "payment_method",
                                    "items"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "description": "Customer's name",
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "qris",
                                            "debit"
                                        ],
                                        "example": "cash"
                                    },
                                    "cash_received": {
                                        "description": "Required if payment_method is cash",
                                        "type": "integer",
                                        "example": 100000
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "Product ID (null for custom items)",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_custom": {
                                                    "description": "Is this a custom item?",
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "custom_name": {
                                                    "description": "Required if is_custom=true",
                                                    "type": "string",
                                                    "example": "Gule Setengah Porsi"
                                                },
                                                "custom_price": {
                                                    "description": "Required if is_custom=true",
                                                    "type": "integer",
                                                    "example": 15000
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "example": "Extra sugar"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction by ID",
                "operationId": "b3d11351fb6ea87ee797d24314e3b7fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transaction ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}